# App.API/Controllers/Timesheet/TimesheetEntryController.ts
import { Request as ExpressRequest } from 'express';
import {
  Body,
  Controller,
  Get,
  Post,
  Put,
  Delete,
  Route,
  Tags,
  Path,
  Security,
  Request,
} from 'tsoa';
import { Service } from 'typedi';

import {
  CreateTimesheetEntryDto,
  TimesheetEntryResponseDto,
  UpdateTimesheetEntryDto,
} from '../../Dtos/Timesheet/TimesheetDto';
import User from '../../Entities/Users/User';
import { TimesheetEntry } from '../../Entities/Timesheets/TimesheetEntry';
import { TimesheetEntryService } from '../../Services/Timesheet/TimesheetEntryService';

/**
 * @summary Controller for managing timesheet entries.
 * @tags Timesheet Entries
 * @security jwt
 */
@Route('timesheet-entries')
@Tags('Timesheet Entries')
@Security('jwt')
@Service()
export class TimesheetEntryController extends Controller {
  constructor(private readonly timesheetEntryService: TimesheetEntryService) {
    super();
  }

  /**
   * @summary Creates a new timesheet entry for the authenticated user.
   * @param dto The data for creating the timesheet entry.
   * @param request The Express request object, containing user information.
   * @returns The newly created timesheet entry.
   * @throws {UnprocessableEntityError} If validation of the DTO fails.
   */
  @Post('/')
  public async createTimesheetEntry(
    @Body() dto: CreateTimesheetEntryDto,
    @Request() request: ExpressRequest,
  ): Promise<TimesheetEntryResponseDto> {
    const me = request.user as User;
    const created = await this.timesheetEntryService.createTimesheetEntry(me.companyId, me.id, dto);
    return this.mapToResponseDto(created);
  }

  /**
   * @summary Retrieves a single timesheet entry by its ID.
   * @param id The ID of the timesheet entry to retrieve.
   * @returns The timesheet entry details.
   * @throws {NotFoundError} If the timesheet entry is not found.
   */
  @Get('/{id}')
  public async getTimesheetEntry(@Path() id: string): Promise<TimesheetEntryResponseDto> {
    const entry = await this.timesheetEntryService.getTimesheetEntryById(id);
    return this.mapToResponseDto(entry);
  }

  /**
   * @summary Updates an existing timesheet entry.
   * @param id The ID of the timesheet entry to update.
   * @param dto The data for updating the timesheet entry.
   * @returns The updated timesheet entry details.
   * @throws {UnprocessableEntityError} If validation of the DTO fails.
   * @throws {NotFoundError} If the timesheet entry to update is not found.
   */
  @Put('/{id}')
  public async updateTimesheetEntry(
    @Path() id: string,
    @Body() dto: UpdateTimesheetEntryDto,
  ): Promise<TimesheetEntryResponseDto> {
    const updated = await this.timesheetEntryService.updateTimesheetEntry(id, dto);
    return this.mapToResponseDto(updated);
  }

  /**
   * @summary Deletes a timesheet entry by its ID.
   * @param id The ID of the timesheet entry to delete.
   * @returns A Promise that resolves upon successful deletion.
   * @throws {NotFoundError} If the timesheet entry to delete is not found.
   */
  @Delete('/{id}')
  public async deleteTimesheetEntry(@Path() id: string): Promise<void> {
    await this.timesheetEntryService.deleteTimesheetEntry(id);
  }

  /**
   * @summary Submits a timesheet entry for approval.
   * @param id The ID of the timesheet entry to submit.
   */
  @Put('/{id}/submit')
  public async submitTimesheetEntry(@Path() id: string): Promise<TimesheetEntryResponseDto> {
    const entry = await this.timesheetEntryService.submitTimesheetEntry(id);
    return this.mapToResponseDto(entry);
  }

  /**
   * @summary Approves a pending timesheet entry.
   * @param id The ID of the timesheet entry to approve.
   */
  @Put('/{id}/approve')
  public async approveTimesheetEntry(@Path() id: string): Promise<TimesheetEntryResponseDto> {
    const entry = await this.timesheetEntryService.approveTimesheetEntry(id);
    return this.mapToResponseDto(entry);
  }

  /**
   * @summary Rejects a pending timesheet entry.
   * @param id The ID of the timesheet entry to reject.
   */
  @Put('/{id}/reject')
  public async rejectTimesheetEntry(@Path() id: string): Promise<TimesheetEntryResponseDto> {
    const entry = await this.timesheetEntryService.rejectTimesheetEntry(id);
    return this.mapToResponseDto(entry);
  }

  /**
   * @summary Marks an approved timesheet entry as invoiced.
   * @param id The ID of the timesheet entry to mark as invoiced.
   */
  @Put('/{id}/invoice')
  public async invoiceTimesheetEntry(@Path() id: string): Promise<TimesheetEntryResponseDto> {
    const entry = await this.timesheetEntryService.invoiceTimesheetEntry(id);
    return this.mapToResponseDto(entry);
  }

  private mapToResponseDto(entry: TimesheetEntry): TimesheetEntryResponseDto {
    return {
      id: entry.id,
      actionCodeId: entry.actionCodeId,
      day: entry.day,
      durationMin: entry.durationMin,
      country: entry.country,
      workMode: entry.workMode,
      note: entry.note,
      status: entry.status,
      statusUpdatedAt: entry.statusUpdatedAt,
    };
  }
}