# App.API/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"TimesheetStatus": {
				"enum": [
					"DRAFT",
					"SUBMITTED",
					"APPROVED",
					"REJECTED"
				],
				"type": "string"
			},
			"TimesheetRowBillableTag": {
				"enum": [
					"billable",
					"non-billable",
					"auto"
				],
				"type": "string"
			},
			"TimesheetRowLocation": {
				"enum": [
					"Office",
					"Homeworking",
					"Hybrid"
				],
				"type": "string"
			},
			"TimesheetRowStatus": {
				"enum": [
					"draft",
					"submitted",
					"approved",
					"rejected"
				],
				"type": "string"
			},
			"TimesheetWeekRowEntryDto": {
				"properties": {
					"day": {
						"type": "string",
						"description": "ISO date string representing the day of the entry."
					},
					"minutes": {
						"type": "number",
						"format": "double",
						"description": "Minutes logged for the day."
					},
					"note": {
						"type": "string",
						"nullable": true,
						"description": "Optional notes for the day."
					}
				},
				"required": [
					"day",
					"minutes"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TimesheetWeekRowRejectionDto": {
				"properties": {
					"reason": {
						"type": "string"
					},
					"actorId": {
						"type": "string"
					},
					"actorName": {
						"type": "string"
					},
					"occurredAt": {
						"type": "string"
					}
				},
				"required": [
					"occurredAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TimesheetWeekRowDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"activityLabel": {
						"type": "string"
					},
					"timeCodeId": {
						"type": "string"
					},
					"billable": {
						"$ref": "#/components/schemas/TimesheetRowBillableTag"
					},
					"location": {
						"$ref": "#/components/schemas/TimesheetRowLocation"
					},
					"countryCode": {
						"type": "string"
					},
					"employeeCountryCode": {
						"type": "string",
						"nullable": true
					},
					"status": {
						"$ref": "#/components/schemas/TimesheetRowStatus"
					},
					"locked": {
						"type": "boolean"
					},
					"entries": {
						"items": {
							"$ref": "#/components/schemas/TimesheetWeekRowEntryDto"
						},
						"type": "array"
					},
					"rejection": {
						"$ref": "#/components/schemas/TimesheetWeekRowRejectionDto"
					}
				},
				"required": [
					"activityLabel",
					"timeCodeId",
					"billable",
					"location",
					"countryCode",
					"status",
					"locked",
					"entries"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TimesheetWeekSettingsDto": {
				"properties": {
					"defaultCountryCode": {
						"type": "string"
					},
					"defaultLocation": {
						"type": "string"
					},
					"maxWeeklyMinutes": {
						"type": "number",
						"format": "double"
					},
					"autoSubmitAt": {
						"type": "string",
						"nullable": true
					},
					"officeCountryCodes": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"maxWeeklyMinutes"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TimesheetWeekRejectionDto": {
				"properties": {
					"reason": {
						"type": "string"
					},
					"actorId": {
						"type": "string"
					},
					"actorName": {
						"type": "string"
					},
					"occurredAt": {
						"type": "string"
					}
				},
				"required": [
					"occurredAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TimesheetWeekResponseDto": {
				"properties": {
					"weekStart": {
						"type": "string"
					},
					"weekEnd": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/TimesheetStatus"
					},
					"rows": {
						"items": {
							"$ref": "#/components/schemas/TimesheetWeekRowDto"
						},
						"type": "array"
					},
					"settings": {
						"$ref": "#/components/schemas/TimesheetWeekSettingsDto"
					},
					"rejection": {
						"$ref": "#/components/schemas/TimesheetWeekRejectionDto"
					}
				},
				"required": [
					"weekStart",
					"weekEnd",
					"status",
					"rows",
					"settings"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TimesheetWeekUpsertDto": {
				"properties": {
					"rows": {
						"items": {
							"$ref": "#/components/schemas/TimesheetWeekRowDto"
						},
						"type": "array"
					}
				},
				"required": [
					"rows"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TimesheetWeekSubmitDto": {
				"properties": {
					"force": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"WorkMode": {
				"enum": [
					"office",
					"remote",
					"hybrid"
				],
				"type": "string"
			},
			"TimesheetEntryStatus": {
				"enum": [
					"saved",
					"pending_approval",
					"approved",
					"rejected",
					"invoiced"
				],
				"type": "string"
			},
...
			"TimesheetEntryResponseDto": {
				"properties": {
					"id": {
						"type": "string",
						"example": "e1n2t3r4-y5i6d7-8901-2345-67890abcdef"
					},
					"actionCodeId": {
						"type": "string",
						"example": "a1c2t3i4-o5n6-7890-1234-567890abcdef"
					},
					"day": {
						"type": "string",
						"example": "2024-01-01"
					},
					"durationMin": {
						"type": "number",
						"format": "double",
						"example": 480
					},
					"country": {
						"type": "string",
						"example": "US"
					},
					"workMode": {
						"$ref": "#/components/schemas/WorkMode",
						"example": "office"
					},
					"note": {
						"type": "string",
						"example": "Worked on feature Y"
					},
					"status": {
						"$ref": "#/components/schemas/TimesheetEntryStatus",
						"example": "pending_approval"
					},
					"statusUpdatedAt": {
						"type": "string",
						"format": "date-time",
						"example": "2024-01-01T15:45:00Z"
					}
				},
				"required": [
					"id",
					"actionCodeId",
					"day",
					"durationMin",
					"country",
					"workMode",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateTimesheetEntryDto": {
				"properties": {
					"actionCodeId": {
						"type": "string",
						"example": "a1c2t3i4-o5n6-7890-1234-567890abcdef"
					},
					"day": {
						"type": "string",
						"example": "2024-01-01"
					},
					"durationMin": {
						"type": "number",
						"format": "double",
						"example": 480
					},
					"country": {
						"type": "string",
						"example": "US"
					},
					"workMode": {
						"$ref": "#/components/schemas/WorkMode",
						"example": "office"
					},
					"note": {
						"type": "string",
						"example": "Worked on feature Y"
					},
					"status": {
						"$ref": "#/components/schemas/TimesheetEntryStatus",
						"example": "saved"
					}
				},
				"required": [
					"actionCodeId",
					"day",
					"durationMin",
					"country"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateTimesheetEntryDto": {
				"properties": {
					"actionCodeId": {
						"type": "string",
						"example": "b1u2g3f4-i5x6-7890-1234-567890abcdef"
					},
					"day": {
						"type": "string",
						"example": "2024-01-02"
					},
					"durationMin": {
						"type": "number",
						"format": "double",
						"example": 540
					},
					"country": {
						"type": "string",
						"example": "CA"
					},
					"workMode": {
						"$ref": "#/components/schemas/WorkMode",
						"example": "remote"
					},
					"note": {
						"type": "string",
						"example": "Continued work on feature Y"
					},
					"status": {
						"$ref": "#/components/schemas/TimesheetEntryStatus",
						"example": "pending_approval"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"TimesheetResponseDto": {
				"properties": {
					"id": {
						"type": "string",
						"example": "t1i2m3e4-s5h6e7e8-9012-3456-7890abcdef"
					},
					"userId": {
						"type": "string",
						"example": "u1s2e3r4-i5d6-7890-1234-567890abcdef"
					},
					"periodStart": {
						"type": "string",
						"example": "2024-01-01"
					},
					"periodEnd": {
						"type": "string",
						"example": "2024-01-07"
					},
					"status": {
						"type": "string",
						"example": "DRAFT"
					},
					"totalMinutes": {
						"type": "number",
						"format": "double",
						"example": 480
					},
					"notes": {
						"type": "string",
						"example": "Weekly report for Project Alpha"
					},
					"entries": {
						"items": {
							"$ref": "#/components/schemas/TimesheetEntryResponseDto"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"userId",
					"periodStart",
					"periodEnd",
					"status",
					"totalMinutes"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateTimesheetDto": {
				"properties": {
					"periodStart": {
						"type": "string",
						"example": "2024-01-01"
					},
					"periodEnd": {
						"type": "string",
						"example": "2024-01-07"
					},
					"notes": {
						"type": "string",
						"example": "Weekly report for project X"
					}
				},
				"required": [
					"periodStart",
					"periodEnd"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateTimesheetDto": {
				"properties": {
					"periodStart": {
						"type": "string",
						"example": "2024-01-01"
					},
					"periodEnd": {
						"type": "string",
						"example": "2024-01-07"
					},
					"notes": {
						"type": "string",
						"example": "Updated weekly report for project X"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "gogotime-api",
		"version": "1.0.0",
		"description": "API for T-DEV-700 Project for Epitech",
		"contact": {
			"name": "Lazaro, Alexy, Massi, Lounis"
		}
	},
	"paths": {
...
		"/timesheet/weeks/{weekStart}/timesheet": {
			"get": {
				"operationId": "GetWeekTimesheet",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TimesheetWeekResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Timesheet Weeks"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "weekStart",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpsertWeekTimesheet",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TimesheetWeekResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Timesheet Weeks"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "weekStart",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TimesheetWeekUpsertDto"
							}
						}
					}
				}
			}
		},
		"/timesheet/weeks/{weekStart}/submit": {
			"post": {
				"operationId": "SubmitWeekTimesheet",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TimesheetWeekResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Timesheet Weeks"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "weekStart",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TimesheetWeekSubmitDto"
							}
						}
					}
				}
			}
		},
...
		"/timesheet-entries": {
			"post": {
				"operationId": "CreateTimesheetEntry",
				"responses": {
					"200": {
						"description": "The newly created timesheet entry.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TimesheetEntryResponseDto"
								}
							}
						}
					}
				},
				"summary": "Creates a new timesheet entry for the authenticated user.",
				"tags": [
					"Timesheet Entries"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "The data for creating the timesheet entry.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTimesheetEntryDto",
								"description": "The data for creating the timesheet entry."
							}
						}
					}
				}
			}
		},
		"/timesheet-entries/{id}": {
			"get": {
				"operationId": "GetTimesheetEntry",
				"responses": {
					"200": {
						"description": "The timesheet entry details.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TimesheetEntryResponseDto"
								}
							}
						}
					}
				},
				"summary": "Retrieves a single timesheet entry by its ID.",
				"tags": [
					"Timesheet Entries"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The ID of the timesheet entry to retrieve.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateTimesheetEntry",
				"responses": {
					"200": {
						"description": "The updated timesheet entry details.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TimesheetEntryResponseDto"
								}
							}
						}
					}
				},
				"summary": "Updates an existing timesheet entry.",
				"tags": [
					"Timesheet Entries"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The ID of the timesheet entry to update.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The data for updating the timesheet entry.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTimesheetEntryDto",
								"description": "The data for updating the timesheet entry."
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteTimesheetEntry",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"summary": "Deletes a timesheet entry by its ID.",
				"tags": [
					"Timesheet Entries"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The ID of the timesheet entry to delete.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/timesheet-entries/{id}/submit": {
			"put": {
				"operationId": "SubmitTimesheetEntry",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TimesheetEntryResponseDto"
								}
							}
						}
					}
				},
				"summary": "Submits a timesheet entry for approval.",
				"tags": [
					"Timesheet Entries"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The ID of the timesheet entry to submit.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/timesheet-entries/{id}/approve": {
			"put": {
				"operationId": "ApproveTimesheetEntry",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TimesheetEntryResponseDto"
								}
							}
						}
					}
				},
				"summary": "Approves a pending timesheet entry.",
				"tags": [
					"Timesheet Entries"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The ID of the timesheet entry to approve.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/timesheet-entries/{id}/reject": {
			"put": {
				"operationId": "RejectTimesheetEntry",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TimesheetEntryResponseDto"
								}
							}
						}
					}
				},
				"summary": "Rejects a pending timesheet entry.",
				"tags": [
					"Timesheet Entries"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The ID of the timesheet entry to reject.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/timesheet-entries/{id}/invoice": {
			"put": {
				"operationId": "InvoiceTimesheetEntry",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TimesheetEntryResponseDto"
								}
							}
						}
					}
				},
				"summary": "Marks an approved timesheet entry as invoiced.",
				"tags": [
					"Timesheet Entries"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The ID of the timesheet entry to mark as invoiced.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/timesheets": {
			"post": {
				"operationId": "CreateTimesheet",
				"responses": {
					"200": {
						"description": "The newly created timesheet.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TimesheetResponseDto"
								}
							}
						}
					}
				},
				"summary": "Creates a new timesheet for the authenticated user.",
				"tags": [
					"Timesheets"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "The data for creating the timesheet.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTimesheetDto",
								"description": "The data for creating the timesheet."
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllTimesheetsForUser",
				"responses": {
					"200": {
						"description": "An array of timesheet details.",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TimesheetResponseDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Retrieves all timesheets for the authenticated user.",
				"tags": [
					"Timesheets"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/timesheets/{id}": {
			"get": {
				"operationId": "GetTimesheet",
				"responses": {
					"200": {
						"description": "The timesheet details.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TimesheetResponseDto"
								}
							}
						}
					}
				},
				"summary": "Retrieves a single timesheet by its ID.",
				"tags": [
					"Timesheets"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The ID of the timesheet to retrieve.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateTimesheet",
				"responses": {
					"200": {
						"description": "The updated timesheet.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TimesheetResponseDto"
								}
							}
						}
					}
				},
				"summary": "Updates an existing timesheet.",
				"tags": [
					"Timesheets"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The ID of the timesheet to update.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The data for updating the timesheet.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTimesheetDto",
								"description": "The data for updating the timesheet."
							}
						}
					}
				}
			}
		},
...
		"/timesheets/{id}/submit": {
			"put": {
				"operationId": "SubmitTimesheet",
				"responses": {
					"200": {
						"description": "The updated timesheet.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TimesheetResponseDto"
								}
							}
						}
					}
				},
				"summary": "Submits a timesheet for approval.",
				"tags": [
					"Timesheets"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The ID of the timesheet to submit.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/timesheets/{id}/approve": {
			"put": {
				"operationId": "ApproveTimesheet",
				"responses": {
					"200": {
						"description": "The updated timesheet.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TimesheetResponseDto"
								}
							}
						}
					}
				},
				"summary": "Approves a submitted timesheet.",
				"tags": [
					"Timesheets"
				],
				"security": [
					{
						"jwt": [
							"manager",
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "The ID of the timesheet to approve.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/timesheets/{id}/reject": {
			"put": {
				"operationId": "RejectTimesheet",
				"responses": {
					"200": {
						"description": "The updated timesheet.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TimesheetResponseDto"
								}
							}
						}
					}
				},
				"summary": "Rejects a submitted timesheet.",
				"tags": [
					"Timesheets"
				],
				"security": [
					{
						"jwt": [
							"manager",
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "The ID of the timesheet to reject.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The body containing the reason for rejection.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"reason": {
										"type": "string"
									}
								},
								"required": [
									"reason"
								],
								"type": "object",
								"description": "The body containing the reason for rejection."
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}