# App.Web/src/lib/api/models/TimesheetRowLocation.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum TimesheetRowLocation {
    OFFICE = 'Office',
    HOMEWORKING = 'Homeworking',
    HYBRID = 'Hybrid',
}