# App.Web/src/lib/api/models/TimesheetWeekSettingsDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TimesheetWeekSettingsDto = {
    defaultCountryCode?: string;
    defaultLocation?: string;
    maxWeeklyMinutes: number;
    autoSubmitAt?: string | null;
    officeCountryCodes?: Array<string>;
};