# App.Web/src/lib/api/models/TimesheetWeekUpsertDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { TimesheetWeekRowDto } from './TimesheetWeekRowDto';
export type TimesheetWeekUpsertDto = {
    rows: Array<TimesheetWeekRowDto>;
};